Localization running

+++ ros2/src/vtr_navigation/include/vtr_navigation/modules/ransac/stereo_ransac_module.hpp
@@ -28,7 +28,7 @@ class StereoRansacModule : public RansacModule {
   /** \brief Collection of config parameters */
   struct Config : RansacModule::Config {
     double mask_depth;
-    double mask_depth_inlier_count;
+    int mask_depth_inlier_count;
     // Whether or not to use landmark covariance for the error model
     bool use_covariance = false;
   };
